+++ district_crop_yield/models/Mid_season_price_prediction.py
@@ -1,8 +1,6 @@
 import pandas as pd
 import numpy as np
 from sklearn.ensemble import RandomForestRegressor
-from sklearn.metrics import mean_squared_error, mean_absolute_error
-
 
 # ---------------------------
 # 1. Data loading & cleaning
