+++ district_crop_yield/utils/utils.py
@@ -9,6 +9,8 @@ from huggingface_hub import login
 from deep_translator import GoogleTranslator
 from models.Mid_season_price_prediction import evaluate_model
 from dotenv import load_dotenv
+from datetime import datetime, date
+import json
 
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__),  '..')))
@@ -299,15 +301,6 @@ def _snake_to_title(s: str) -> str:
     s = re.sub(r"([a-z])([A-Z])", r"\1 \2", s)    # camelCase -> spaced
     return s.strip().title()
 
-def _prettify_value(v):
-    if isinstance(v, bool):
-        return "Yes" if v else "No"
-    # numbers that look like money: you can define heuristic keys instead of autos
-    if isinstance(v, (int, float)) and not isinstance(v, bool):
-        # don't automatically format all numbers — but often money fields benefit:
-        return _fmt_amt(v)
-    return v
-
 def prettify_details(details, rename_map=None, drop_keys=None, money_keys=None, keep_raw=False):
     """
     - details: dict-like or object (will use vars()).
@@ -368,3 +361,78 @@ def prettify_details(details, rename_map=None, drop_keys=None, money_keys=None, 
         pretty["Raw Details"] = d
 
     return pretty
+
+
+
+def to_serializable(val):
+    """Convert common non-JSON types to JSON-serializable Python primitives."""
+    # None
+    if val is None:
+        return None
+
+    # Numpy scalars
+    if isinstance(val, (np.generic,)):
+        return val.item()
+
+    # Numpy arrays -> lists
+    if isinstance(val, (np.ndarray,)):
+        return [to_serializable(x) for x in val.tolist()]
+
+    # Built-in collections
+    if isinstance(val, (list, tuple, set)):
+        return [to_serializable(x) for x in val]
+
+    if isinstance(val, dict):
+        return {str(k): to_serializable(v) for k, v in val.items()}
+
+    # Decimal -> float or str (Decimal may be large-precision)
+    if isinstance(val, Decimal):
+        return float(val)
+
+    # datetimes
+    if isinstance(val, (datetime, date)):
+        return val.isoformat()
+
+    # floats with Inf/NaN (JSON doesn't support Infinity/NaN)
+    if isinstance(val, float):
+        if math.isinf(val) or math.isnan(val):
+            # Choose how to map these — I use None, you can use "Infinity" string if you prefer
+            return None
+        return val
+
+    # ints, bools, str are fine
+    if isinstance(val, (int, bool, str)):
+        return val
+
+    # custom objects: attempt to serialize __dict__
+    if hasattr(val, "__dict__"):
+        return to_serializable(vars(val))
+
+    # fallback: try repr (safe) OR string
+    try:
+        return str(val)
+    except Exception:
+        return None
+
+def serialize_recommendations(recs):
+    """Serialize recommendations or any nested structure to JSON-safe Python types."""
+    return to_serializable(recs)
+
+def debug_json(obj):
+    try:
+        return json.dumps(obj)
+    except TypeError as e:
+        # print problematic types (walk small sample)
+        def walk(x, path="root"):
+            if isinstance(x, dict):
+                for k,v in x.items():
+                    walk(v, f"{path}.{k}")
+            elif isinstance(x, (list, tuple, set)):
+                for i, v in enumerate(x):
+                    walk(v, f"{path}[{i}]")
+            else:
+                if not isinstance(x, (str, int, float, bool, type(None))):
+                    print("Non-serializable found at", path, "type:", type(x), "value:", repr(x))
+        walk(obj)
+        raise
+
